InventoreFragment and Navigation Graph implementation

+++ app/src/main/java/com/example/inventory/data/item/Item.kt
@@ -8,6 +8,6 @@ import androidx.room.PrimaryKey
 data class Item (
     @PrimaryKey @ColumnInfo(name = "id")val id: Int,
     @ColumnInfo(name = "name") val name: String,
-    @ColumnInfo(name = "price") val price: Int,
+    @ColumnInfo(name = "price") val price: Double,
     @ColumnInfo(name = "quantity") val quantity: Int
 )
